+++ plot/line.go
@@ -43,6 +43,7 @@ type LineChartConfig struct {
 
 // CreateLineChart generates and returns a *charts.Line object based on LineChartConfig.
 func CreateLineChart(config LineChartConfig, data ...insyra.IDataList) *charts.Line {
+	data = nonNilLists("CreateLineChart", data)
 	if len(data) == 0 {
 		insyra.LogWarning("plot", "CreateLineChart", "No data available for line chart. Returning nil.")
 		return nil